Adding testing framework
Adding a small and simple testing framework that will help fight
regression errors. The tests are located in the tests directory
and are invoked by running this command:

node scheme.js -t

Note the -t turns the tests on and they are automatically run. In
a later feature I will make it so all tests in the tests folder
are run by that command. Only hard coded tests are run.